Refactored code and fixed issue on switches
